+++ flame/trafgen.h
@@ -11,6 +11,9 @@
 #include "query.h"
 #include "tcpsession.h"
 
+#include "quicly.h"
+#include "quicly/streambuf.h"
+
 #include <TokenBucket.h>
 #include <uvw.hpp>
 
@@ -45,6 +48,9 @@ class TrafGen
     std::shared_ptr<uvw::TcpHandle> _tcp_handle;
     std::shared_ptr<TCPSession> _tcp_session;
 
+    quicly_context_t ctx;
+    quicly_cid_plaintext_t next_cid;
+
     std::shared_ptr<uvw::TimerHandle> _sender_timer;
     std::shared_ptr<uvw::TimerHandle> _timeout_timer;
     std::shared_ptr<uvw::TimerHandle> _shutdown_timer;
@@ -66,6 +72,12 @@ class TrafGen
 
     void start_tcp_session();
     void start_wait_timer_for_tcp_finish();
+
+    int q_run_client(int fd, const char *host, struct sockaddr *sa, socklen_t salen);
+    int q_run_loop(int fd, quicly_conn_t *conn, int (*stdin_read_cb)(quicly_conn_t *conn));
+    int q_send_one(int fd, quicly_datagram_t *p);
+    void q_process_msg(quicly_conn_t **conn, struct msghdr *msg, size_t dgram_len);
+    int q_read_stdin(quicly_conn_t *conn);
 
 public:
     TrafGen(std::shared_ptr<uvw::Loop> l,
